requirements.txt
streamlit
numpy
requests
faiss-cpu
sentence-transformers
PyPDF2
python-dotenv